Made enhancements as per the comments
