+++ pkg/aws/aws.go
@@ -616,47 +616,3 @@ func getTime(rawReason string) (time.Time, error) {
 
 	return extractedTime, nil
 }
-
-func BlockEgress(ctx context.Context, ec2Client EC2API, securityGroupID string) error {
-	input := &ec2v2.RevokeSecurityGroupEgressInput{
-		GroupId: &securityGroupID,
-		IpPermissions: []ec2v2types.IpPermission{
-			{
-				IpProtocol: awsString("-1"), // -1 = all protocols
-				IpRanges: []ec2v2types.IpRange{
-					{CidrIp: awsString("0.0.0.0/0")},
-				},
-			},
-		},
-	}
-
-	_, err := ec2Client.RevokeSecurityGroupEgress(ctx, input)
-	if err != nil {
-		return fmt.Errorf("failed to revoke egress: %w", err)
-	}
-	return nil
-}
-
-func RestoreEgress(ctx context.Context, ec2Client EC2API, securityGroupID string) error {
-	input := &ec2v2.AuthorizeSecurityGroupEgressInput{
-		GroupId: &securityGroupID,
-		IpPermissions: []ec2v2types.IpPermission{
-			{
-				IpProtocol: awsString("-1"),
-				IpRanges: []ec2v2types.IpRange{
-					{CidrIp: awsString("0.0.0.0/0")},
-				},
-			},
-		},
-	}
-
-	_, err := ec2Client.AuthorizeSecurityGroupEgress(ctx, input)
-	if err != nil {
-		return fmt.Errorf("failed to restore egress: %w", err)
-	}
-	return nil
-}
-
-func awsString(value string) *string {
-	return &value
-}
